refactor: remove compatibility shims and barrel re-exports
Delete polyfills, supabase client barrel, icons barrel, and index.ts
re-export files. Import hooks, services, and lucide icons directly.
Remove unused CSS shimmer token block and restore HomeRoute/TournamentFlow
fixes from incomplete merges.
