+++ config/vitest.config.ts
@@ -10,7 +10,7 @@ export default defineConfig({
 			"@/app": path.resolve(__dirname, "..", "src/app"),
 			"@/features": path.resolve(__dirname, "..", "src/features"),
 			"@/shared": path.resolve(__dirname, "..", "src/shared"),
-			"@supabase/client": path.resolve(__dirname, "..", "src/shared/services/supabase/client.ts"),
+			"@supabase/client": path.resolve(__dirname, "..", "src/shared/services/supabase/runtime.ts"),
 			"@supabase/types": path.resolve(__dirname, "..", "supabase/types.ts"),
 			"@db": path.resolve(__dirname, "..", "supabase"),
 		},
