+++ src/app/AppShell.tsx
@@ -1,7 +1,7 @@
 import { lazy, Suspense, useLayoutEffect } from "react";
 import { Navigate, Route, Routes, useLocation } from "react-router-dom";
 import { AppLayout } from "@/shared/components/layout/AppLayout";
-import { Loading } from "@/shared/components/layout/Feedback";
+import { Loading } from "@/shared/components/layout/Feedback/Loading";
 
 const HomeRoute = lazy(() => import("@/app/routes/HomeRoute"));
 const AdminRoute = lazy(() => import("@/app/routes/AdminRoute"));
