+++ src/features/admin/components/AdminNamesTab.tsx
@@ -1,7 +1,7 @@
 import type { ChangeEvent } from "react";
 import Button from "@/shared/components/layout/Button";
 import { Input } from "@/shared/components/layout/FormPrimitives";
-import { Eye, EyeOff, Loader2, Lock, Trash2 } from "@/shared/lib/icons";
+import { Eye, EyeOff, Loader2, Lock, Trash2 } from "lucide-react";
 import { isNameHidden, isNameLocked } from "@/shared/lib/names/nameFilters";
 import type { NameItem } from "@/shared/types";
