+++ src/features/analytics/RankingAdjustment.tsx
@@ -10,7 +10,7 @@ import {
 import { Button, CardBody, CardHeader, Chip, cn, Divider } from "@heroui/react";
 import { motion } from "framer-motion";
 import { memo, useEffect, useRef, useState } from "react";
-import { GripVertical, Loader2, Save } from "@/shared/lib/icons";
+import { GripVertical, Loader2, Save } from "lucide-react";
 import { ErrorManager } from "@/shared/services/errorManager";
 import type { NameItem } from "@/shared/types";
