+++ src/features/analytics/components/HiddenNamesPanel.tsx
@@ -1,6 +1,6 @@
 import Button from "@/shared/components/layout/Button";
 import { EmptyState } from "@/shared/components/layout/EmptyState";
-import { Eye, EyeOff } from "@/shared/lib/icons";
+import { Eye, EyeOff } from "lucide-react";
 import { Panel, SectionHeader } from "./DashboardPrimitives";
 
 interface HiddenName {
