+++ src/features/analytics/components/LeaderboardPanel.tsx
@@ -1,7 +1,7 @@
 import Button from "@/shared/components/layout/Button";
 import { EmptyState } from "@/shared/components/layout/EmptyState";
-import { Loading } from "@/shared/components/layout/Feedback";
-import { Trophy } from "@/shared/lib/icons";
+import { Loading } from "@/shared/components/layout/Feedback/Loading";
+import { Trophy } from "lucide-react";
 import { ContextBadge, Panel, SectionHeader } from "./DashboardPrimitives";
 
 interface LeaderboardEntry {
