+++ src/features/analytics/components/RecentActivityPanel.tsx
@@ -1,5 +1,5 @@
 import Button from "@/shared/components/layout/Button";
-import { Activity, TrendingUp, Trophy, Users } from "@/shared/lib/icons";
+import { Activity, TrendingUp, Trophy, Users } from "lucide-react";
 import type { DashboardTimeframe } from "../hooks/useDashboardData";
 import { Panel, SectionHeader, StatTile } from "./DashboardPrimitives";
