+++ src/features/tournament/Tournament.tsx
@@ -1,7 +1,7 @@
 import { AnimatePresence, motion, useReducedMotion } from "framer-motion";
 import { type KeyboardEvent, memo, useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { useNavigate } from "react-router-dom";
-import { ErrorComponent } from "@/shared/components/layout/Feedback";
+import { ErrorComponent } from "@/shared/components/layout/Feedback/ErrorBoundary";
 import { CAT_IMAGES } from "@/shared/lib/constants";
 import { getRandomCatImage } from "@/shared/lib/media";
 import { getVisibleNames } from "@/shared/lib/names/nameFilters";
@@ -11,7 +11,7 @@ import { MatchSideCard } from "./components/MatchSideCard";
 import { TournamentAnnouncements } from "./components/TournamentAnnouncements";
 import { TournamentComplete } from "./components/TournamentComplete";
 import { TournamentHeader } from "./components/TournamentHeader";
-import { useAudioManager } from "./hooks";
+import { useAudioManager } from "./hooks/useHelpers";
 import { useTournamentState } from "./hooks/useTournamentState";
 import { getHeatLevel, type HeatLevel, STREAK_THRESHOLDS } from "./utils/heat";
 import { extractMatchData, getMatchSideId } from "./utils/matchHelpers";
