--- src/features/tournament/hooks/index.ts
@@ -1,4 +1,0 @@
-export { useAudioManager } from "./useHelpers";
-export { useNameSuggestion } from "./useNameSuggestion";
-export { useTournamentHandlers } from "./useTournamentHandlers";
-export { useSaveTournamentRatings } from "./useSaveTournamentRatings";
