+++ src/shared/components/layout/AppLayout.tsx
@@ -2,12 +2,9 @@ import { Analytics } from "@vercel/analytics/react";
 import { lazy, type ReactNode, Suspense } from "react";
 import { shouldEnableAnalytics } from "@/app/analytics";
 import { ScrollToTopButton } from "@/shared/components/layout/Button";
-import {
-	ErrorBoundary,
-	ErrorComponent,
-	Loading,
-	OfflineIndicator,
-} from "@/shared/components/layout/Feedback";
+import { ErrorBoundary, ErrorComponent } from "@/shared/components/layout/Feedback/ErrorBoundary";
+import { Loading } from "@/shared/components/layout/Feedback/Loading";
+import { OfflineIndicator } from "@/shared/components/layout/Feedback/OfflineIndicator";
 import { FloatingNavbar } from "@/shared/components/layout/FloatingNavbar";
 import useAppStore from "@/store/appStore";
