+++ src/shared/components/layout/Button.tsx
@@ -1,7 +1,7 @@
 import { cva } from "class-variance-authority";
 import React, { memo } from "react";
 import { cn } from "@/shared/lib/utils";
-import { ChevronDown, Loader2 } from "@/shared/lib/icons";
+import { ChevronDown, Loader2 } from "lucide-react";
 import "./FancyButton.css";
 
 /**
