--- src/shared/components/layout/Feedback/index.ts
@@ -1,6 +1,0 @@
-export {
-	ErrorBoundary,
-	ErrorComponent,
-} from "./ErrorBoundary";
-export { Loading } from "./Loading";
-export { OfflineIndicator } from "./OfflineIndicator";
