--- src/shared/components/ui/index.ts
@@ -1,5 +1,0 @@
-export { CinematicFooter } from "./cinematic-footer";
-export { default as CinematicFooterDemo } from "./cinematic-footer-demo";
-export { CinematicHero } from "./cinematic-landing-hero";
-export { default as CinematicHeroDemo } from "./cinematic-landing-hero-demo";
-export { CinematicHero as CinematicHeroAdapted } from "./CinematicHero";
