--- src/shared/lib/icons.ts
@@ -1,43 +1,0 @@
-export {
-        Activity,
-        BarChart3,
-        Cat,
-        Check,
-        CheckCircle,
-        ChevronDown,
-        ChevronLeft,
-        ChevronRight,
-        Clock,
-        Copy,
-        Eye,
-        EyeOff,
-        Gamepad2,
-        GripVertical,
-        Heart,
-        Layers,
-        LayoutGrid,
-        Lightbulb,
-        Loader2,
-        Lock,
-        LogOut,
-        Medal,
-        Music,
-        PawPrint,
-        Pencil,
-        Plus,
-        Save,
-        SkipBack,
-        SkipForward,
-        Target,
-        Trash2,
-        TrendingUp,
-        Trophy,
-        Undo2,
-        User,
-        Users,
-        Volume2,
-        VolumeX,
-        X,
-        ZoomIn,
-        ZoomOut,
-} from "lucide-react";
