--- src/shared/services/supabase/client.ts
@@ -1,4 +1,0 @@
-export * from "./imageService";
-export * from "./ratingService";
-export * from "./statsService";
-export * from "./runtime";
